requirements.txt
regex
# used in compiling C++ function
pybind11
# from markdown_it import MarkdownIt
markdown-it-py
# from bs4 import BeautifulSoup, Tag, NavigableString
beautifulsoup4

# extra packages
mysql-connector

